Fix: can't delete default category

+++ src/modules/categories/categories.service.ts
@@ -137,7 +137,7 @@ export const deleteCategoryService = async (
   });
 
   if (!category) throw new Error('Category not found');
-  // if (category.isDefault) throw new Error('Cannot delete default categories');
+  if (category.isDefault) throw new Error('Cannot delete default categories');
 
   await prisma.transactionCategory.delete({
     where: { id: categoryId }
